[DEV-84] refactor: 코드 피드백, 버그 반영
