+++ src/hooks/query/use-todo.ts
@@ -1,12 +1,15 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 
 import { invalidateTodoRelatedQueries } from "@/services/invalidate";
+import { todoKeys } from "@/services/query-key";
 import {
   createTodo,
   deleteTodo,
+  getTodoDetail,
   TodoParams,
   updateTodo,
 } from "@/services/todo";
+import { TodoResponse } from "@/types/todo";
 
 export const useCreateTodo = () => {
   const queryClient = useQueryClient();
@@ -44,5 +47,16 @@ export const useDeleteTodo = (goalId?: number) => {
     onSuccess: () => {
       invalidateTodoRelatedQueries(queryClient, goalId);
     },
+  });
+};
+
+export const useTodoDetail = (todoId: number, isEditMode: boolean) => {
+  return useQuery<TodoResponse>({
+    queryKey: todoKeys.detail(todoId),
+    queryFn: () => {
+      if (!todoId) throw new Error("todoId가 필요합니다.");
+      return getTodoDetail(todoId);
+    },
+    enabled: !isEditMode && Boolean(todoId),
   });
 };
