+++ src/styles/globals.css
@@ -59,3 +59,8 @@ body {
   list-style-type: disc;
   padding-left: 1.5rem;
 }
+
+h1,
+h2 {
+  border-bottom: none !important;
+}
